+++ src/frontend/apps/impress/src/features/docs/doc-export/mappingPDF.tsx
@@ -2,6 +2,7 @@ import { pdfDefaultSchemaMappings } from '@blocknote/xl-pdf-exporter';
 
 import {
   blockMappingCalloutPDF,
+  blockMappingDatabasePDF,
   blockMappingDividerPDF,
   blockMappingHeadingPDF,
   blockMappingImagePDF,
@@ -16,6 +17,7 @@ export const pdfDocsSchemaMappings: DocsExporterPDF['mappings'] = {
   blockMapping: {
     ...pdfDefaultSchemaMappings.blockMapping,
     callout: blockMappingCalloutPDF,
+    database: blockMappingDatabasePDF, // TODO: create Db block
     heading: blockMappingHeadingPDF,
     image: blockMappingImagePDF,
     paragraph: blockMappingParagraphPDF,